requirements.txt
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
discord.py==1.5.1